fix: fix post/pre decrement/increment addressing

+++ sirc-vm/peripheral-cpu/src/execution/fetch_and_decode.rs
@@ -140,7 +140,7 @@ pub fn decode_and_register_fetch(
     let instruction_type =
         decode_fetch_and_decode_step_instruction_type(implied_representation.op_code);
 
-    let addr_inc: i8 = match implied_representation.op_code {
+    let addr_inc: i16 = match implied_representation.op_code {
         Instruction::LoadRegisterFromIndirectRegisterPostIncrement => 1, // TODO: Match LOAD (a)+
         Instruction::StoreRegisterToIndirectRegisterPreDecrement => -1,  // TODO: Match STOR -(a)
         _ => 0,
